Move all hooks to hooks folder
