+++ packages/root-cause-core/lib/hooks/errorInStepHook.ts
@@ -1,0 +1,13 @@
+import { AfterHook, AfterHookArgs } from '../interfaces';
+import { unknownErrorToOurRepresentation } from './hookUtils';
+
+export const errorInStepHook: AfterHook = async function errorInStepHook({
+  instrumentedFunctionResult,
+  testContext,
+}: AfterHookArgs) {
+  if (!instrumentedFunctionResult.success) {
+    testContext.addStepMetadata({
+      stepError: unknownErrorToOurRepresentation(instrumentedFunctionResult.error),
+    });
+  }
+};
