+++ packages/root-cause-core/lib/hooks/networkLogsHooks.ts
@@ -1,10 +1,10 @@
 import type { Page as PuppeteerPage, CDPSession } from 'puppeteer';
 import type { Page as PlaywrightPage, CDPSession as PlaywrightCDPSession } from 'playwright';
-import type { TestContext } from './TestContext';
-import type { BeforeAllHook, AfterAllHook } from './interfaces';
-import { isNotPlaywrightPage, isPlaywrightChromiumBrowserContext } from './utils';
+import type { TestContext } from '../TestContext';
+import type { BeforeAllHook, AfterAllHook } from '../interfaces';
+import { isNotPlaywrightPage, isPlaywrightChromiumBrowserContext } from '../utils';
 import chromeHar, { ChromeHarMessage } from 'chrome-har';
-import { addDisposer, runAllDisposers } from './hooksHandlersDisposersHelper';
+import { addDisposer, runAllDisposers } from '../hooksHandlersDisposersHelper';
 import fs from 'fs-extra';
 import path from 'path';
