+++ packages/root-cause-core/lib/hooks/screenshotCollectionHook.ts
