+++ packages/root-cause-core/lib/hooks/stacktraceHook.ts
@@ -1,5 +1,5 @@
-import { extractCodeLocationDetailsSync } from './utils';
-import { BeforeHook } from './interfaces';
+import { extractCodeLocationDetailsSync } from '../utils';
+import { BeforeHook } from '../interfaces';
 
 export const stacktraceHook: BeforeHook = async function stacktraceHook({ testContext }) {
   const stepCodeLocation = extractCodeLocationDetailsSync(testContext.testFilePath, process.cwd());
